added improvements

+++ Backend/src/services/productServices/deleteProduct.ts
@@ -4,6 +4,18 @@ const prisma = new PrismaClient();
 
 async function deleteProducts(ids: number[]) {
   try {
+
+    const isExisting = await prisma.product.findMany({
+      where: {
+        productId: { in: ids }
+      }
+
+    });
+
+    if (isExisting.length === 0) {
+      logger.error("No products found with the provided IDs", { ids });
+      return new Error("No products found with the provided IDs");
+    }
     const { count } = await prisma.product.deleteMany({
       where: {
         productId: { in: ids }
